+++ lib/src/views/chats/chat_messages_components/chat_list_entry.dart
@@ -10,6 +10,7 @@ import 'package:twonly/src/database/tables/messages.table.dart'
 import 'package:twonly/src/database/twonly.db.dart';
 import 'package:twonly/src/model/memory_item.model.dart';
 import 'package:twonly/src/services/mediafiles/mediafile.service.dart';
+import 'package:twonly/src/utils/log.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/chat_reaction_row.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_audio_entry.dart';
 import 'package:twonly/src/views/chats/chat_messages_components/entries/chat_contacts.entry.dart';
@@ -79,6 +80,10 @@ class _ChatListEntryState extends State<ChatListEntry> {
         if (mediaFiles != null) {
           mediaService = MediaFileService(mediaFiles);
           if (mounted) setState(() {});
+        } else {
+          Log.error(
+            'Media file not found for ${widget.message.messageId} => ${widget.message.mediaId}',
+          );
         }
       });
     }